Move the `calculateMD5` function into its own file
This allows us to remove a closure, and we also change the code to initialize various constants lazily.
