+++ test/unit/crypto_spec.js
@@ -17,7 +17,6 @@ import {
   AES128Cipher,
   AES256Cipher,
   ARCFourCipher,
-  calculateMD5,
   calculateSHA256,
   calculateSHA384,
   calculateSHA512,
@@ -31,6 +30,7 @@ import {
   PasswordResponses,
   stringToBytes,
 } from "../../src/shared/util.js";
+import { calculateMD5 } from "../../src/core/calculate_md5.js";
 
 describe("crypto", function () {
   function hex2binary(s) {
